add template to store listeners
also extract both listeners to a single class `StoreUnavailableEventListener`.
