+++ app/Resources/views/Pages/store-disabled.html.twig
@@ -1,0 +1,6 @@
+{% extends '::_base.html.twig' %}
+
+
+{% block body %}
+    <h1>Store disabled</h1>
+{% endblock %}
