+++ app/Resources/views/Pages/store-under-construction.html.twig
@@ -1,0 +1,6 @@
+{% extends '::_base.html.twig' %}
+
+
+{% block body %}
+    <h1>Store under construction</h1>
+{% endblock %}
