+++ src/Elcodi/Store/CoreBundle/Resources/config/classes.yml
@@ -3,5 +3,5 @@ parameters:
     #
     # Event Listeners
     #
-    store.core.event_listener.store_disabled.class: Elcodi\Store\CoreBundle\EventListener\StoreDisabledEventListener
-    store.core.event_listener.store_under_construction.class: Elcodi\Store\CoreBundle\EventListener\StoreUnderConstructionEventListener
+    store.core.event_listener.store_disabled.class: Elcodi\Store\CoreBundle\EventListener\StoreUnavailableEventListener
+    store.core.event_listener.store_under_construction.class: Elcodi\Store\CoreBundle\EventListener\StoreUnavailableEventListener
